Make handler work with draft GC
During a draft GC all drafts older than a specific time are deleted.
It is therefore not possible to recalculate elements for a specific draft and no draft ID can be obtained from the CQN.
We just skip our custom before handler in that case, which is anyway mainly useful in UI scenarios.