Don't mix destination configuration with SAP API Business Hub API Key
Reading the API Key is explicitly implemented in this sample application.
Adding that configuration to the destination configuration of a Remote Service creates the impression that this is natively supported by CAP.
This is however not the case. It also promotes the bad behaviour of extending CAPs configuration schemas with custom properties.