fix: Fix compiler info messages to make the project message-free (#245)
info messages are hidden by default. If enabled, there is a message
about doc-comments not having any effect. `/** */` comments have special
meaning in CDS: They contain documentation that is assigned to an
artifact as a `doc` property in CSN. To help users identify places where
such a comment has no effect, an info message is provided.

This is just to be 100% message free.