doc(README): Make shell snippet copy&paste safe
Many bash/shell snippets in the README use an environment variable
`YOUR_CONTAINER_REGISTRY`, which is used as `$YOUR_CONTAINER_REGISTRY`.

However, that has the downside that if users copy & paste the snippet
and accidentally press enter or use a shell that doesn't allow multi-
line pasting (and executes it immediately), then the variable will be
replaced by an empty string, because it wasn't set before.

It's good practice to make those snippets copy&paste-safe.
We can achieve that by adding `:?` to the variable, which emits an
error if the variable is unset or null.

See https://devhints.io/bash for a bash cheat sheet.

Example:

```sh
docker push ${YOUR_CONTAINER_REGISTRY:?}/bookshop-srv
```

On ZSH:

```
zsh: YOUR_CONTAINER_REGISTRY: parameter not set
```

On Bash:

```
bash: YOUR_CONTAINER_REGISTRY: parameter null or not set
```