Upgrade all components to cds8 (#345)
This PR fixes test failures of the Bookshop ITests using CAP 3 and cds8.

* `authentication.normalize-provider-tenant: true` needs to be set in
the `application.yaml` because current itest logic requires the provider
tenant to be subscribed. The deployment shell script is setting this
value to `false` using a `sed` command which requires this property to
be available in the file
* `hdi-deploy` in version `5` does not bundle any hana db client, but
comes with a post install script that checks that one client has been
installed by an external party. As we don´t need db accesses in the
ìnstall-dependencies` goal in `pom.xml`, we just set `--ignore-scripts`

---------

Co-authored-by: Marc Becker <marc.becker@sap.com>